refactor(model): 패키지 이동에 따라 generator 클래스 패키지명 수정
LottoGenerator, DefaultLottoGenerator
- Lotto 클래스를 ticket 패키지로 이동함에 따라 의존하고 있는 클래스들의 경로명 수정
