+++ src/main/java/lotto/model/generator/DefaultLottoGenerator.java
@@ -3,7 +3,7 @@ package lotto.model.generator;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
-import lotto.model.Lotto;
+import lotto.model.ticket.Lotto;
 
 public class DefaultLottoGenerator implements LottoGenerator {
