+++ src/main/java/lotto/model/generator/LottoGenerator.java
@@ -1,7 +1,7 @@
 package lotto.model.generator;
 
 import java.util.List;
-import lotto.model.Lotto;
+import lotto.model.ticket.Lotto;
 
 public interface LottoGenerator {
